+++ src/Symfony/Component/Form/Tests/FormFactoryTest.php
@@ -446,13 +446,13 @@ class FormFactoryTest extends \PHPUnit_Framework_TestCase
     public function testOptionsCanBeOverridden()
     {
         $this->guesser1->expects($this->once())
-                ->method('guessType')
-                ->with('Application\Author', 'firstName')
-                ->will($this->returnValue(new TypeGuess(
-                    'text',
-                    array('attr' => array('maxlength' => 10)),
-                    Guess::MEDIUM_CONFIDENCE
-                )));
+            ->method('guessType')
+            ->with('Application\Author', 'firstName')
+            ->will($this->returnValue(new TypeGuess(
+                'text',
+                array('attr' => array('maxlength' => 10)),
+                Guess::MEDIUM_CONFIDENCE
+            )));
 
         $factory = $this->getMockFactory(array('createNamedBuilder'));
 
@@ -474,20 +474,20 @@ class FormFactoryTest extends \PHPUnit_Framework_TestCase
     public function testCreateBuilderUsesMaxLengthIfFound()
     {
         $this->guesser1->expects($this->once())
-                ->method('guessMaxLength')
-                ->with('Application\Author', 'firstName')
-                ->will($this->returnValue(new ValueGuess(
-                    15,
-                    Guess::MEDIUM_CONFIDENCE
-                )));
+            ->method('guessAttributes')
+            ->with('Application\Author', 'firstName')
+            ->will($this->returnValue(array('maxlength' => new ValueGuess(
+                15,
+                Guess::MEDIUM_CONFIDENCE
+            ))));
 
         $this->guesser2->expects($this->once())
-                ->method('guessMaxLength')
-                ->with('Application\Author', 'firstName')
-                ->will($this->returnValue(new ValueGuess(
-                    20,
-                    Guess::HIGH_CONFIDENCE
-                )));
+            ->method('guessAttributes')
+            ->with('Application\Author', 'firstName')
+            ->will($this->returnValue(array('maxlength' => new ValueGuess(
+                20,
+                Guess::HIGH_CONFIDENCE
+            ))));
 
         $factory = $this->getMockFactory(array('createNamedBuilder'));
 
@@ -504,23 +504,126 @@ class FormFactoryTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('builderInstance', $this->builder);
     }
 
+    public function testCreateBuilderUsesMinAndMaxValueIfFound()
+    {
+        $this->guesser1->expects($this->once())
+            ->method('guessType')
+            ->will($this->returnValue(new TypeGuess(
+                'integer',
+                array(),
+                Guess::HIGH_CONFIDENCE
+            )));
+
+        $this->guesser1->expects($this->once())
+            ->method('guessAttributes')
+            ->with('Application\Temperature', 'degrees')
+            ->will($this->returnValue(array('min' => new ValueGuess(
+                -276,
+                Guess::HIGH_CONFIDENCE
+            ))));
+
+        $this->guesser2->expects($this->once())
+            ->method('guessAttributes')
+            ->with('Application\Temperature', 'degrees')
+            ->will($this->returnValue(array('max' => new ValueGuess(
+                100,
+                Guess::HIGH_CONFIDENCE
+            ))));
+
+        $factory = $this->getMockFactory(array('createNamedBuilder'));
+
+        $factory->expects($this->once())
+            ->method('createNamedBuilder')
+            ->with('degrees', 'integer', null, array('attr' => array('min' => -276, 'max' => 100)))
+            ->will($this->returnValue('builderInstance'));
+
+        $this->builder = $factory->createBuilderForProperty(
+            'Application\Temperature',
+            'degrees'
+        );
+
+        $this->assertEquals('builderInstance', $this->builder);
+    }
+
+    public function testMinAndMaxAttributesCanBeOverridden()
+    {
+        $this->guesser1->expects($this->once())
+            ->method('guessType')
+            ->will($this->returnValue(new TypeGuess(
+                'integer',
+                array(),
+                Guess::HIGH_CONFIDENCE
+            )));
+
+        $this->guesser1->expects($this->once())
+            ->method('guessAttributes')
+            ->with('Application\Temperature', 'degrees')
+            ->will($this->returnValue(array('min' => new ValueGuess(
+                -276,
+                Guess::HIGH_CONFIDENCE
+            ))));
+
+        $factory = $this->getMockFactory(array('createNamedBuilder'));
+
+        $factory->expects($this->once())
+            ->method('createNamedBuilder')
+            ->with('degrees', 'integer', null, array('attr' => array('min' => 50)))
+            ->will($this->returnValue('builderInstance'));
+
+        $this->builder = $factory->createBuilderForProperty(
+            'Application\Temperature',
+            'degrees',
+            null,
+            array('attr' => array('min' => 50))
+        );
+
+        $this->assertEquals('builderInstance', $this->builder);
+    }
+
+    public function testMinAndMaxAttributesAreNotAddedToTextType()
+    {
+        $this->guesser1->expects($this->once())
+            ->method('guessAttributes')
+            ->with('Application\Author', 'firstName')
+            ->will($this->returnValue(array('min' => new ValueGuess(
+                -276,
+                Guess::HIGH_CONFIDENCE
+            ))));
+
+        $factory = $this->getMockFactory(array('createNamedBuilder'));
+
+        $factory->expects($this->once())
+            ->method('createNamedBuilder')
+            ->with('firstName', 'text', null)
+            ->will($this->returnValue('builderInstance'));
+
+        $this->builder = $factory->createBuilderForProperty(
+            'Application\Author',
+            'firstName',
+            null,
+            array('attr' => array('min' => 50))
+        );
+
+        $this->assertEquals('builderInstance', $this->builder);
+    }
+
     public function testCreateBuilderUsesRequiredSettingWithHighestConfidence()
     {
         $this->guesser1->expects($this->once())
-                ->method('guessRequired')
-                ->with('Application\Author', 'firstName')
-                ->will($this->returnValue(new ValueGuess(
-                    true,
-                    Guess::MEDIUM_CONFIDENCE
-                )));
+            ->method('guessRequired')
+            ->with('Application\Author', 'firstName')
+            ->will($this->returnValue(new ValueGuess(
+                true,
+                Guess::MEDIUM_CONFIDENCE
+            )));
 
         $this->guesser2->expects($this->once())
-                ->method('guessRequired')
-                ->with('Application\Author', 'firstName')
-                ->will($this->returnValue(new ValueGuess(
-                    false,
-                    Guess::HIGH_CONFIDENCE
-                )));
+            ->method('guessRequired')
+            ->with('Application\Author', 'firstName')
+            ->will($this->returnValue(new ValueGuess(
+                false,
+                Guess::HIGH_CONFIDENCE
+            )));
 
         $factory = $this->getMockFactory(array('createNamedBuilder'));
 
@@ -540,26 +643,51 @@ class FormFactoryTest extends \PHPUnit_Framework_TestCase
     public function testCreateBuilderUsesPatternIfFound()
     {
         $this->guesser1->expects($this->once())
-                ->method('guessPattern')
-                ->with('Application\Author', 'firstName')
-                ->will($this->returnValue(new ValueGuess(
-                    '[a-z]',
-                    Guess::MEDIUM_CONFIDENCE
-                )));
+            ->method('guessAttributes')
+            ->with('Application\Author', 'firstName')
+            ->will($this->returnValue(array('pattern' => new ValueGuess(
+                '[a-z]',
+                Guess::MEDIUM_CONFIDENCE
+            ))));
 
         $this->guesser2->expects($this->once())
-                ->method('guessPattern')
-                ->with('Application\Author', 'firstName')
-                ->will($this->returnValue(new ValueGuess(
-                    '[a-zA-Z]',
-                    Guess::HIGH_CONFIDENCE
-                )));
+            ->method('guessAttributes')
+            ->with('Application\Author', 'firstName')
+            ->will($this->returnValue(array('pattern' => new ValueGuess(
+                '[a-zA-Z]',
+                Guess::HIGH_CONFIDENCE
+            ))));
 
         $factory = $this->getMockFactory(array('createNamedBuilder'));
 
         $factory->expects($this->once())
             ->method('createNamedBuilder')
             ->with('firstName', 'text', null, array('attr' => array('pattern' => '[a-zA-Z]')))
+            ->will($this->returnValue('builderInstance'));
+
+        $this->builder = $factory->createBuilderForProperty(
+            'Application\Author',
+            'firstName'
+        );
+
+        $this->assertEquals('builderInstance', $this->builder);
+    }
+
+    public function testCreateBuilderWithNullAttribute()
+    {
+        $this->guesser1->expects($this->once())
+            ->method('guessAttributes')
+            ->with('Application\Author', 'firstName')
+            ->will($this->returnValue(array('maxlength' => new ValueGuess(
+                null,
+                Guess::MEDIUM_CONFIDENCE
+            ))));
+
+        $factory = $this->getMockFactory(array('createNamedBuilder'));
+
+        $factory->expects($this->once())
+            ->method('createNamedBuilder')
+            ->with('firstName', 'text', null, array())
             ->will($this->returnValue('builderInstance'));
 
         $this->builder = $factory->createBuilderForProperty(